// AngularMain/src/app/modules/models/FinancialService/TransactionHdDto.ts
export interface TransactionHdDto{
    tenentId: number;
    mytransid: number;
    locationId: number;
    employeeId: number;
    serviceId: number;
    masterServiceId: number;
    serviceType: number;
    serviceSubType: number;
    source: string;
    attachId: number;
    transDocNo: string;
    bankId: number;
    chequeNumber: string;
    chequeDate: Date;
    chequeAmount: number;
    totinstallments: number;
    totamt: number;
    amtPaid: number;
    loanAct: string;
    hajjAct: string;
    persLoanAct: string;
    consumerLoanAct: string;
    otherAct1: string;
    otherAct2: string;
    otherAct3: string;
    otherAct4: string;
    otherAct5: string;
    serApproval1: string;
    approvalBy1: string;
    approvedDate1: Date;
    serApproval2: string;
    approvalBy2: string;
    approvedDate2: Date;
    serApproval3: string;
    approvalBy3: string;
    approvedDate3: Date;
    serApproval4: string;
    approvalBy4: string;
    approvedDate4: Date;
    serApproval5: string;
    approvalBy5: string;
    approvedDate5: Date;
    activitycode: string;
    mydocno: number;
    userbatchno: string;
    projectno: string;
    transdate: Date;
    reference: string;
    notes: string;
    glpostref: string;
    glpostref1: string;
    companyid: number;
    discount: number;
    terms: number;
    refTransId: number;
    signatures: string;
    extraSwitch1: string;
    extraSwitch2: string;
    status: string;
    crupId: number;
    userid: string;
    active: boolean;
    entrydate: Date;
    entrytime: Date;
    updttime: Date;
    installmentAmount:number;
    installmentsBegDate: Date;
    untilMonth: string;
    documentAttachmentForm: any[];
    discountType:number;
    eachInstallmentsAmt:number;
    allowDiscountDefault:boolean;
}